requirements.txt
pygame
numpy
pillow